seperate desktop stream for plan-b
screenshare track can now be send as a seperate video track. the key info is the videoType

Most of the logic control based on videoType information.

Please read the comments in the source files for details

depends on sdp-simulcast which is modified for additional desktop track ssrc and msid which are written into the video mLine.